# My_Warehouse/api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateRemainder:
    properties:
      amount:
        type: integer
      barcode:
        type: string
      branch_id:
        type: string
      categroy_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  models.UpdateRemainder:
    properties:
      amount:
        type: integer
      barcode:
        type: string
      branch_id:
        type: string
      categroy_id:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
paths:
...
  /remainder:
    get:
      consumes:
      - application/json
      description: Get List Remainder
      operationId: get_list_remainder
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search_by_filial
        in: query
        name: SearchByFilial
        type: string
      - description: search_by_category
        in: query
        name: SearchByCategory
        type: string
      - description: search_by_barcode
        in: query
        name: SearchByBarcode
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Remainder
      tags:
      - Remainder
    post:
      consumes:
      - application/json
      description: Create Remainder
      operationId: create_remainder
      parameters:
      - description: CreateRemainderRequest
        in: body
        name: remainder
        required: true
        schema:
          $ref: '#/definitions/models.CreateRemainder'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Remainder
      tags:
      - Remainder
  /remainder/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Remainder
      operationId: delete_remainder
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Remainder
      tags:
      - Remainder
    get:
      consumes:
      - application/json
      description: Get By ID Remainder
      operationId: get_by_id_remainder
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Remainder
      tags:
      - Remainder
    put:
      consumes:
      - application/json
      description: Update Remainder
      operationId: update_remainder
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateRemainderRequest
        in: body
        name: remainder
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRemainder'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Remainder
      tags:
      - Remainder
swagger: "2.0"